vsock: consolidate integration tests, remove < 1.12 support
